Heartbeats no longer use the vhost's inbox
