+++ host_core/lib/application.ex
@@ -101,6 +101,7 @@ defmodule HostCore.Application do
       {Phoenix.PubSub, name: :hostcore_pubsub},
       {Task.Supervisor, name: ControlInterfaceTaskSupervisor},
       {Task.Supervisor, name: HostTaskSupervisor},
+      {Task.Supervisor, name: ProviderTaskSupervisor},
       {Task.Supervisor, name: InvocationTaskSupervisor},
       {HostCore.Actors.ActorRpcSupervisor, strategy: :one_for_one},
       {HostCore.Providers.ProviderSupervisor, strategy: :one_for_one, name: ProviderRoot},
