+++ host_core/lib/host_core/vhost/heartbeats.ex
@@ -4,10 +4,41 @@ defmodule HostCore.Vhost.Heartbeats do
   virtual host from which the heartbeat eminates.
   """
 
+  @thirty_seconds 30_000
+
+  use GenServer, restart: :transient
+
   alias HostCore.Actors.ActorSupervisor
   alias HostCore.CloudEvent
   alias HostCore.Providers.ProviderSupervisor
   alias Timex.Format.Duration.Formatters.Humanized
+
+  def start_link(host_pid, host_key) when is_pid(host_pid) do
+    GenServer.start_link(__MODULE__, {host_pid, host_key}, name: String.to_atom("hb-#{host_key}"))
+  end
+
+  @impl true
+  def init({host_pid, host_key}) do
+    :timer.send_interval(@thirty_seconds, self(), :publish_heartbeat)
+
+    {:ok, {host_pid, host_key}}
+  end
+
+  @impl true
+  def handle_info(:publish_heartbeat, state) do
+    publish_heartbeat(state)
+    {:noreply, state}
+  end
+
+  defp publish_heartbeat({host_pid, _host_key}) do
+    if Process.alive?(host_pid) do
+      state = HostCore.Vhost.VirtualHost.full_state(host_pid)
+
+      state
+      |> generate_heartbeat()
+      |> Enum.each(fn hb -> CloudEvent.publish(hb, state.config.lattice_prefix) end)
+    end
+  end
 
   def generate_heartbeat(state) do
     config = state.config
